chore: shifting unzip middle to backend-common
