+++ backend/backend-common/package.json
@@ -20,11 +20,14 @@
   "dependencies": {
     "@paybox/common": "*",
     "@paybox/zeus": "^0.1.2",
+    "@types/pako": "^2.0.3",
     "bcryptjs": "^2.4.3",
+    "compression": "^1.7.4",
     "crypto": "^1.0.1",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "jose": "^5.2.0",
+    "pako": "^2.1.0",
     "redis": "^4.6.12",
     "uuid": "^9.0.1"
   },
